Established GPT API
Still need to figure out how to integrate it into html but I have gotten the API access and ensured it works. There are some tests ya'll can do as well if you want it to make sure it works on your system. Refer to the comments on the top to get started with it